Restructuring whole system of how SVN detects the current revision a file is on, to work around the bug of it not being able to detect "MOVE" revisions.  Essentially the bug (maybe in SVN) is that the second revision in the output of an svn status -v USUALLY tells you the last revision in which this file was actually "changed".  Well, maybe that is a matter of opinion, but I consider a Move, and serious change as a file's presence or lack-of-presence in a given directory will pretty much screw up any codebase...
Anyhow.  svn status -v ignores move operations for that second
revision operation.  So for the sake of a fix, I have to go the
long way around and do an svn LOG (much slower) and compare from
the first revision from svn status -v.  It gets me there, and is
reliable, but slow.  So...  To compensate, I added a global svn
log caching system.  At the moment I haven't added the CREATE
statement for SQLite, but, just ask if you are upgrading.